talent_gateway/requirements.txt
appdirs==1.4.2
Django==1.10.6
django-allauth==0.31.0
django-datatables-view==1.14.0
django-json-field==0.5.7
oauthlib==2.0.1
packaging==16.8
psycopg2==2.7
pyparsing==2.2.0
python-dateutil==2.6.0
python-openid==2.2.5
requests==2.13.0
requests-oauthlib==0.8.0
simplejson==3.10.0
six==1.10.0
South==1.0.2
openpyxl==2.4.5
pyunpack==0.1.2
rarfile==3.0
sqlalchemy==1.1.9
django-celery-results==1.0.1
supervisor==3.3.1
django-bootstrap-form==3.2.1
uwsgi==2.0.15
boto==2.46.1
